Rename 2021 Climate Report.pdf to 2021 DEI Climate Report.pdf
This initiative assessed DEI-related issues experienced by graduate students within the University of Arizona School of Sociology in order to provide recommendations for addressing inequities. Data was collected via a survey incorporating both closed and open-ended questions, which was administered to 54 doctoral students using Qualtrics, and 9 focus groups with 24 total participants. Results indicated that 60%+ of graduate students had experienced or witness instances of hostility or discrimination, 75%+ had experienced financial insecurity, and that 30% experienced a lack of support with caregiving responsibilities. Findings were distributed to School of Sociology leadership in the form of both a formal report and oral presentation, and provided concrete action items to address DEI issues and improve equity.